Refactor damage calculation logic to separate module
