+++ entities/player.py
@@ -2,7 +2,7 @@
 
 class Player:
     def __init__(self, x, y): 
-        self.position = (x, y)
         self.hp = 10
         self.weapon = 0
         self.armor = 0
+
